Customized some things
One important change is to the relative-line-numbers face. For some reason, the
solarized themes make the line numbers inherit properties from the text on their
line - if the text on the line was colored blue, the line number would be blue
as well.

I fixed this by making the relative-line-numbers face inherit its properties
from both the linum *and* default faces. Per standard, it only inherits from linum.