Current line number highlight (bg) now only enabled in active window
To achieve this, I had to do some pretty dirty things - namely, creating my own
window focus in & out hooks, which are in no way reliable. Currently, they only
trigger when other-window or one of the windmove functions is called. I would
like to expand it further, though, as there isn't any such hook built into
Emacs, to my knowledge.

Overall, I'm quite happy with the result. Yes, it's overkill for such an
incredibly simple feature, but I'm glad it now works the way I want it. And
despite the (somewhat) ugly implementation, it works very reliably!

There are, however, some additional cases where I'd like the hook to trigger.
For instance, when activating the minibuffer. I just googled this and it turns
out there are built-in (hooray!) hooks for this purpose: minibuffer-setup-hook
and minibuffer-exit-hook, respectively. That is next on my todo list.