toggle-line-numbers: background now works (almost) perfectly
The only problem I have found occurs when loading a new theme. To update the
current line number background then, I'd have to loop through all buffers, which
is something that I'm not 100% sure about how to do. For that reason, and for
the fact that I quite seldom switch themes in a session, it is a low-priority
issue.

Otherwise, it now works quite perfectly and should emulate hl-line's behavior
perfectly, but probably more efficiently, as I am not using 'post-command-hook
for everything.