Documented bug; made org-tree-view/toggle-visibility reversible
It's completely reversible in that the position of the cursor will return to the
original headline after reversing the command (i.e., running it again). It works
by not moving the cursor at all :-) Sounds simple, but it's tricky when you
can't just move to the old `point', because the points may have changed.